+++ src/main/java/net/jrodolfo/java_evolution/java01/serialization/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for Java 1.1 object serialization.
+ * Executable examples for Java 1.1 object serialization.
  */
 package net.jrodolfo.java_evolution.java01.serialization;